Fix compile errors, backport input code for RR
# Conflicts:
#	platform/Windows/voidsw.vcxproj
#	source/blood/src/controls.cpp
#	source/blood/src/m32common.cpp
#	source/rr/src/actors.cpp
#	source/rr/src/anim.cpp
#	source/rr/src/demo.cpp
#	source/rr/src/game.cpp
#	source/rr/src/gameexec.cpp
#	source/rr/src/global.h
#	source/rr/src/menus.cpp
#	source/rr/src/menus.h
#	source/rr/src/net.cpp
#	source/rr/src/player.cpp
#	source/rr/src/premap.cpp
#	source/rr/src/savegame.cpp
#	source/rr/src/sbar.cpp
#	source/rr/src/screens.cpp
#	source/rr/src/sector.cpp

+++ source/rr/src/screens.cpp
@@ -273,12 +273,12 @@ static void G_DrawCameraText(int16_t i)
         rotatesprite_win(22<<16, 163<<16, 65536L, 512, CAMCORNER+1, 0, 0, 2+4);
         rotatesprite_win((310-10)<<16, 163<<16, 65536L, 512, CAMCORNER+1, 0, 0, 2);
 
-        if ((int32_t)totalclock & 16)
+        if ((int32_t) totalclock&16)
             rotatesprite_win(46<<16, 32<<16, 65536L, 0, CAMLIGHT, 0, 0, 2);
     }
     else
     {
-        int32_t flipbits = ((int32_t)totalclock << 1) & 48;
+        int32_t flipbits = ((int32_t) totalclock<<1)&48;
 
         for (bssize_t x=-64; x<394; x+=64)
             for (bssize_t y=0; y<200; y+=64)
@@ -584,7 +584,7 @@ static void G_DrawOverheadMap(int32_t cposx, int32_t cposy, int32_t czoom, int16
         if (p == screenpeek || GTFLAGS(GAMETYPE_OTHERPLAYERSINMAP))
         {
             if (pSprite->xvel > 16 && pPlayer->on_ground)
-                i = APLAYERTOP + (((int32_t)totalclock >> 4) & 3);
+                i = APLAYERTOP+(((int32_t) totalclock>>4)&3);
             else
                 i = APLAYERTOP;
 
@@ -1367,8 +1367,8 @@ void G_FadePalette(int32_t r, int32_t g, int32_t b, int32_t e)
     videoFadePalette(r, g, b, e);
     videoNextPage();
 
-    int32_t tc = (int32_t)totalclock;
-    while ((int32_t)totalclock < tc + 4)
+    int32_t tc = (int32_t) totalclock;
+    while (totalclock < tc + 4)
         G_HandleAsync();
 }
 
@@ -1676,7 +1676,7 @@ void G_DisplayLogo(void)
                         soundanm++;
                         S_PlaySound(PIPEBOMB_EXPLODE);
                     }
-                    rotatesprite_fs(160 << 16, 104 << 16, ((int32_t)totalclock - 120) << 10, 0, DUKENUKEM, 0, 0, 2 + 8);
+                    rotatesprite_fs(160<<16, 104<<16, ((int32_t) totalclock-120)<<10, 0, DUKENUKEM, 0, 0, 2+8);
                 }
                 else if (totalclock >= (120+60))
                     rotatesprite_fs(160<<16, (104)<<16, 60<<10, 0, DUKENUKEM, 0, 0, 2+8);
@@ -1690,7 +1690,7 @@ void G_DisplayLogo(void)
                     }
 
                     rotatesprite_fs(160<<16, (104)<<16, 60<<10, 0, DUKENUKEM, 0, 0, 2+8);
-                    rotatesprite_fs(160 << 16, (129) << 16, ((int32_t)totalclock - 220) << 11, 0, THREEDEE, 0, 0, 2 + 8);
+                    rotatesprite_fs(160<<16, (129)<<16, ((int32_t) totalclock - 220)<<11, 0, THREEDEE, 0, 0, 2+8);
                 }
                 else if (totalclock >= (220+30))
                     rotatesprite_fs(160<<16, (129)<<16, 30<<11, 0, THREEDEE, 0, 0, 2+8);
@@ -1700,7 +1700,7 @@ void G_DisplayLogo(void)
                     // JBF 20030804
                     if (totalclock >= 280 && totalclock < 395)
                     {
-                        rotatesprite_fs(160<<16, (151)<<16, (410-(int32_t)totalclock)<<12, 0, PLUTOPAKSPRITE+1, (sintable[((int32_t)totalclock<<4)&2047]>>11), 0, 2+8);
+                        rotatesprite_fs(160<<16, (151)<<16, (410-(int32_t) totalclock)<<12, 0, PLUTOPAKSPRITE+1, (sintable[((int32_t) totalclock<<4)&2047]>>11), 0, 2+8);
                         if (soundanm == 2)
                         {
                             soundanm++;
@@ -1714,7 +1714,7 @@ void G_DisplayLogo(void)
                             soundanm++;
                             S_PlaySound(PIPEBOMB_EXPLODE);
                         }
-                        rotatesprite_fs(160<<16, (151)<<16, 30<<11, 0, PLUTOPAKSPRITE+1, (sintable[((int32_t)totalclock<<4)&2047]>>11), 0, 2+8);
+                        rotatesprite_fs(160<<16, (151)<<16, 30<<11, 0, PLUTOPAKSPRITE+1, (sintable[((int32_t) totalclock<<4)&2047]>>11), 0, 2+8);
                     }
                 }
 
@@ -2383,7 +2383,7 @@ void G_BonusScreen(int32_t bonusonly)
 
                 if (totalclock >= 1000000000 && totalclock < 1000000320)
                 {
-                    switch (((int32_t)totalclock >> 4) % 15)
+                    switch (((int32_t) totalclock>>4)%15)
                     {
                     case 0:
                         if (bonuscnt == 6)
@@ -2423,7 +2423,7 @@ void G_BonusScreen(int32_t bonusonly)
                 else if ((int32_t)totalclock > (10240+120L)) break;
                 else
                 {
-                    switch (((int32_t)totalclock >> 5) & 3)
+                    switch (((int32_t) totalclock>>5)&3)
                     {
                     case 1:
                     case 3:
@@ -2977,7 +2977,7 @@ void G_BonusScreenRRRA(int32_t bonusonly)
 
                 if (totalclock >= 1000000000 && totalclock < 1000000320)
                 {
-                    switch (((uint32_t)totalclock>>4)%15)
+                    switch (((int32_t) totalclock>>4)%15)
                     {
                     case 0:
                         if (bonuscnt == 6)
@@ -3013,7 +3013,7 @@ void G_BonusScreenRRRA(int32_t bonusonly)
                 else if ((int32_t)totalclock > (10240+120L)) break;
                 else
                 {
-                    switch (((int32_t)totalclock >> 5) & 3)
+                    switch (((int32_t) totalclock>>5)&3)
                     {
                     case 1:
                     case 3:
@@ -3205,7 +3205,7 @@ void G_BonusScreenRRRA(int32_t bonusonly)
         P_SetGamePalette(g_player[myconnectindex].ps, BASEPAL, 8+2+1);
         while (1)
         {
-            switch (((int32_t)totalclock >> 4) & 1)
+            switch (((int32_t) totalclock >> 4) & 1)
             {
             case 0:
                 rotatesprite(0,0,65536,0,RRTILE8677,0,0,2+8+16+64+128,0,0,xdim-1,ydim-1);
