+++ source/sw/src/sounds.cpp
@@ -278,7 +278,7 @@ void CheckSndData(char *file, int line)
 //
 
 void
-SoundCallBack(unsigned int num)
+SoundCallBack(intptr_t num)
 {
     VOC_INFOp vp;
